removed JSON from App
